Codebase reorganization, preparation for cyclical tasks enabling
