+++ app/services/budget_task_services.py
@@ -1,0 +1,4 @@
+def prepare_next_month_data():
+    prev_budget = 0
+    if not prev_budget:
+        return 0
