+++ main.py
@@ -1,5 +1,7 @@
 from app import create_app
 
+app, celery = create_app()
+
 if __name__ == "__main__":
-    app = create_app()
+
     app.run(debug=True)
