+++ tests/test_tasks.py
@@ -1,0 +1,8 @@
+from app.tasks.budget_tasks import create_next_month_budget
+
+
+def test_create_next_month_budget():
+    """
+    TODO Write proper test case for when task is ready to create budgets based on incomes and expesnes
+    TODO Write proper test case to test the periodic job - for when the app will be up and running at all times
+    """
